% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_text.R
\name{check_text}
\alias{check_text}
\title{Number of documents per patient counting}
\usage{
check_text(data, plot = FALSE)
}
\arguments{
\item{data}{the name of the data table}

\item{plot}{if TRUE, allows to return the result as an histogram}
}
\value{

}
\description{
Count the number of documents (texts) per patient and return a table or a plot representing the result. This works with a data table previously imported with the “import_data” function.
}
