% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easier_words.R
\name{easier_words}
\alias{easier_words}
\title{Simplify a words list}
\usage{
easier_words(word_list, data = NULL, stopword = FALSE)
}
\arguments{
\item{word_list}{the list of searched words imported with the “import_words” function}

\item{data}{data imported with the “import_data” function (option allowing to collect fuzzy words)}

\item{stopword}{if TRUE, allows to delete stopwords for the fuzzy words research (option)}
}
\value{

}
\description{
Simplify a words list imported with the “import_words” function. The simplification concerns the “grouped words” which are put into an only word instead of being a list of several grouped words which have a difficult structure composed of nested lists.
The function can also allow to collect fuzzy words.
}
