% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Data importation}
\usage{
import_data(
  file,
  IDp_column,
  TEXTE_column,
  IDt_column = NULL,
  date_column = NULL,
  rest = FALSE,
  csv2 = FALSE,
  header = TRUE
)
}
\arguments{
\item{file}{the file pathway}

\item{IDp_column}{an integer indicating the column containing patients ID}

\item{TEXTE_column}{an integer indicating the column containing texts}

\item{IDt_column}{an integer indicating the column containing texts ID}

\item{date_column}{an integer indicating the column containing dates}

\item{rest}{if TRUE, the function add the remaining columns}

\item{csv2}{if TRUE, allows to use csv2 function instead of csv function}

\item{header}{TRUE if the file has a header}
}
\value{

}
\description{
Import data of a csv file as a table and identify the contents of the columns of interest (patients ID, text, text ID, …).
}
