% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_words.R
\name{import_words}
\alias{import_words}
\title{Words importation}
\usage{
import_words(file, simplify = FALSE)
}
\arguments{
\item{file}{the file pathway}

\item{simplify}{if TRUE, allows to directly apply the “easier_words” function (option)}
}
\value{

}
\description{
Import the wanted words written in a csv or pdf file build with a particular semantics.
}
