% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_words.R
\name{make_words}
\alias{make_words}
\title{Build a words list}
\usage{
make_words(mywords, precision = "specific")
}
\arguments{
\item{mywords}{a vector in which we can write the wanted words}

\item{precision}{“specific” or “global” (put “specific” or “global” for the precision mode of each word. It’s then used as a tool to know if we do a fuzzy analysis on the word)}
}
\value{

}
\description{
Build a words list
}
