% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_read.R
\name{manual_read}
\alias{manual_read}
\title{manual_read}
\usage{
manual_read(
  data,
  units_track,
  word_col,
  word,
  test = 1,
  limit = 150,
  sbs = FALSE,
  myid = NULL
)
}
\arguments{
\item{data}{table of data imported with the “import_data” function}

\item{units_track}{table of “tracking values” made with the “units_tracking” function}

\item{word_col}{number of the column corresponding to the wanted key word}

\item{word}{key word (regex used to read the texts quicker)}

\item{test}{if 1, check only cases without NA values, if 2, check only cases with NA values, if 3, check all cases}

\item{limit}{number of documents to check (it will be less if there are less documents to check than the limit)}

\item{sbs}{if TRUE, also to do the analysis document by document, else, all extracted values and code lines to execute are printed}

\item{myid}{a vector allowing to put the wanted text IDs to check. If NULL, all text IDs found in the data are used.}
}
\value{

}
\description{
Allows to do manual verifications (by reading concerned texts) about value extractions.
}
