% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nombres.R
\name{nombres}
\alias{nombres}
\title{nombres with letters}
\usage{
nombres(regex = FALSE, simplify = TRUE)
}
\arguments{
\item{regex}{if TRUE, return a result adapted for regex}

\item{simplify}{if TRUE, remove accents}
}
\value{

}
\description{
Return a vector of numbers (in letters) from 0 to 1000
}
