% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nombriemes.R
\name{nombriemes}
\alias{nombriemes}
\title{nombriemes with letters}
\usage{
nombriemes(regex = FALSE, simplify = TRUE)
}
\arguments{
\item{regex}{if TRUE, return a result adapted for regex}

\item{simplify}{if TRUE, remove accents}
}
\value{

}
\description{
Return a vector of nombriemes (in letters) from 0ieme to 1000eme.
}
