% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_word_here.R
\name{one_word_here}
\alias{one_word_here}
\title{Detect a word in a list of documents}
\usage{
one_word_here(text_list, word, count = FALSE)
}
\arguments{
\item{text_list}{a vector containing all texts to analyse}

\item{word}{the word we want to search in the texts}

\item{count}{if TRUE, allows to count the searched word in each document instead of only indicate if it’s present or not}
}
\value{

}
\description{
For a given word, the function searches this word in all given texts.
}
