% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_tracking.R
\name{patient_tracking}
\alias{patient_tracking}
\title{Make a tracking of values of interest for patients}
\usage{
patient_tracking(units_track, patient = NULL, show_na = FALSE)
}
\arguments{
\item{units_track}{table of “tracking values” made with the “units_tracking” function}

\item{patient}{vector containing patients ID}

\item{show_na}{if TRUE, patients and units with no value are still written but with a “X”, instead of being deleted of the result}
}
\value{

}
\description{
Allows to regroup into a list the organized results of value tracking for a list of wanted patients.
}
