% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_text.R
\name{units_text}
\alias{units_text}
\title{Extraction of text samples of interest}
\usage{
units_text(data, word_list, test = "extract", IDt = FALSE)
}
\arguments{
\item{data}{table of data imported with the “import_data” function}

\item{word_list}{a list of words imported with the “import_words” function or made with the “make_words” function}

\item{test}{take as value either “extract”, “extract_mat”, “extract_char” or “locate”, allowing to have a result respectively as a list, a matrix or a string of characters, or to locate the searched words into the texts (localization done thanks to 2 numbers showing the beginning and the end of the word)}

\item{IDt}{if TRUE, allows to give results with the text ID instead of the text row numbers}
}
\value{

}
\description{
The main goal is to extract text samples containing values of interest. The research is done with words having an indicated “unit type”. This is done from the data table imported with the “import_data” function and from the words list imported with the “import_words” function or made with the “make_words” function.
}
