% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_values.R
\name{units_values}
\alias{units_values}
\title{Extraction of values of interest}
\usage{
units_values(data, word_list, IDt = FALSE)
}
\arguments{
\item{data}{table of data imported with the “import_data” function}

\item{word_list}{a list of words imported with the “import_words” function or made with the “make_words” function}

\item{IDt}{if TRUE, allows to give results with the text ID instead of the text row numbers}
}
\value{

}
\description{
Allows to extract values of interest into the texts. The research is done with words having an indicated “unit type”. This is done from the data table imported with the “import_data” function and from the words list imported with the “import_words” function or made with the “make_words” function. The result is a list of words containing a list of values (integer).
}
