% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where_is_word.R
\name{where_is_word}
\alias{where_is_word}
\title{Detection of documents containing searched words}
\usage{
where_is_word(data, word_list, group = TRUE, IDt = FALSE)
}
\arguments{
\item{data}{table of data imported with the “import_data” function}

\item{word_list}{a list of words imported with the “import_words” function or made with the “make_words” function}

\item{group}{if TRUE, allows to group the “grouped words” into one result instead of giving a result to each one}

\item{IDt}{if TRUE, allows to give results with the text ID instead of the text row numbers}
}
\value{

}
\description{
Allows for each word to give the list of documents containing this word. This is done from the data table imported with the “import_data” function and from the words list imported with the “import_words” function or made with the “make_words” function.
}
