% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_docanalysis.R
\name{word_docanalysis}
\alias{word_docanalysis}
\title{All document analyses at once}
\usage{
word_docanalysis(data, word_list, rank = FALSE)
}
\arguments{
\item{data}{table of data imported with the “import_data” function}

\item{word_list}{a list of words imported with the “import_words” function or made with the “make_words” function}

\item{rank}{if TRUE, allows to return all results in descending order}
}
\value{

}
\description{
Allows to do all document analyses at once, including the “word_hm”, “word_hm_dp” and “word_hm_p” functions. This is done from the data table imported with the “import_data” function and from the words list imported with the “import_words” function or made with the “make_words” function.
}
