% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_hm.R
\name{word_hm}
\alias{word_hm}
\title{Number of documents having searched words}
\usage{
word_hm(text_list, word_list, rank = FALSE, count = FALSE, plot = FALSE)
}
\arguments{
\item{text_list}{a vector containing all texts to analyse}

\item{word_list}{a list of words imported with the “import_words” function or made with the “make_words” function}

\item{rank}{if TRUE, allows to return all results in descending order}

\item{count}{if TRUE, allows to count the searched words in each document instead of only indicate if it’s present or not}

\item{plot}{if TRUE, allows to return results as an histogram}
}
\value{

}
\description{
Count for each word how many documents have this word. The words must be as a list previously imported with the ‘import_words” function or made with the “make_words” function.
}
