% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_hm_dp.R
\name{word_hm_dp}
\alias{word_hm_dp}
\title{Number of documents having searched words per patient}
\usage{
word_hm_dp(data, word_list, rank = FALSE, count = FALSE)
}
\arguments{
\item{data}{table of data imported with the “import_data” function}

\item{word_list}{a list of words imported with the “import_words” function or made with the “make_words” function}

\item{rank}{if TRUE, allows to return all results in descending order}

\item{count}{if TRUE, allows to count the searched words in each document instead of only indicate if it’s present or not}
}
\value{

}
\description{
Count for each word and each patient how many documents of this patient have this word. The words must be as a list previously imported with the ‘import_words” function or made with the “make_words” function.
}
