% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_hm_p.R
\name{word_hm_p}
\alias{word_hm_p}
\title{Number of patients concerned about searched words}
\usage{
word_hm_p(data, word_list, rank = FALSE, plot = FALSE)
}
\arguments{
\item{data}{table of data imported with the “import_data” function}

\item{word_list}{a list of words imported with the “import_words” function or made with the “make_words” function}

\item{rank}{if TRUE, allows to return all results in descending order}

\item{plot}{if TRUE, allows to return results as an histogram}
}
\value{

}
\description{
Count for each word how many patients have at least one document containing this word. The words must be as a list previously imported with the ‘import_words” function or made with the “make_words” function.
}
